VXDTF2: changed all selectionVariables for Two-, Three- and FourHitFilters from float to double following the suggestions of Tobias and Martin Ritter, since float seems to cause non-deterministic behavior sometimes. Tests were adapted to work with the changed interface and other classes using these selectionVariables
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@20217 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/trackFindingVXD/TrackletFilters/include/ZiggZaggXY.h
@@ -33,7 +33,7 @@ namespace Belle2 {
     {
       if (hitContainer.size() < 4) return 1;
 
-      typedef SelVarHelper<PointType, float> Helper;
+      typedef SelVarHelper<PointType, double> Helper;
 
       std::vector<int> chargeSigns;
       chargeSigns.reserve(hitContainer.size() - 2);
